Moved news items from /news to /blog
We are placing all news/blog items in the /blog directory, which is
natively supported by Docusaurus.

This commit includes blog post with an external_url in the metadata.  That
is not currently supported, and we'll need to either support it, or update
blog post to include text and a link to the original item.